Fix max components wrapper
Max components used in decomposition during filtration were
not mapping correctly and always a single component has been used.
Now, some components are found, and adaptive feature filtration
works.

+++ src/Spectre.Algorithms/Parameterization/DivikOptions.cs
@@ -179,7 +179,7 @@ namespace Spectre.Algorithms.Parameterization
 			addParam("PlotRecursively", PlottingRecursively);
 			addParam("DecompositionPlots", PlottingDecomposition);
 			addParam("DecompositionPlotsRecursively", PlottingDecompositionRecursively);
-			addParam("MaxComponentsForDecomposition", MaxComponentsForDecomposition);
+			addParam("MaxComponentsForDecomposition", (double)MaxComponentsForDecomposition);
 			addParam("OutPath", OutputPath);
 			addParam("CachePath", CachePath);
 			addParam("Cache", Caching);
